Update README.md
BrokeBucks is for all the college students and young adults struggling with the financial woes of modern education and living. 

It's a coin that celebrates the hustle, the ramen nights, and the dream of financial freedom.

From ramen noodles to real gains, BrokeBucks is here to turn your financial drought into a comedy of success. 

Every transaction with BrokeBucks is a step towards a future where your wallet isn't just for show. 

Join the BrokeBucks revolution, where we're all in this together, dreaming of the day we're no longer broke, but making bucks!